Moved styles to frontend root
