+++ src/frontend/styles/components/OptionDetailsView.module.scss
