+++ src/frontend/styles/components/OptionListItemView.module.scss
