+++ src/frontend/styles/components/OptionViewContainer.module.scss
