+++ src/pages/index.tsx
@@ -1,5 +1,5 @@
 // @ts-ignore
-import classes from "../styles/pages/LandingPage.module.scss";
+import classes from "../frontend/styles/pages/LandingPage.module.scss";
 
 import Image from "next/image";
 import Layout from "../components/Layout";
